+++ import_export_extensions/api/views/__init__.py
@@ -1,8 +1,12 @@
 from .export_job import (
+    BaseExportJobForUserViewSet,
+    BaseExportJobViewSet,
     ExportJobForUserViewSet,
     ExportJobViewSet,
 )
 from .import_job import (
+    BaseImportJobForUserViewSet,
+    BaseImportJobViewSet,
     ImportJobForUserViewSet,
     ImportJobViewSet,
 )