+++ invocations/docs.py
@@ -10,5 +10,7 @@ LOCAL_DOCS_DIR = pathlib.Path("docs/_build")
 def build(context: invoke.Context):
     """Build documentation."""
     saritasa_invocations.print_success("Start building of local documentation")
-    context.run(f"sphinx-build -E -a docs {LOCAL_DOCS_DIR}")
+    context.run(
+        f"sphinx-build -E -a docs {LOCAL_DOCS_DIR} --exception-on-warning",
+    )
     saritasa_invocations.print_success("Building completed")