+++ test_project/tests/integration_tests/test_api/test_export.py
@@ -11,12 +11,26 @@ from import_export_extensions.models import ExportJob
 
 
 @pytest.mark.django_db(transaction=True)
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            reverse("export-artist-start"),
+            id="Export url",
+        ),
+        pytest.param(
+            f"{reverse('artists-export')}",
+            id="Action url",
+        ),
+    ],
+)
 def test_export_api_creates_export_job(
     admin_api_client: test.APIClient,
+    export_url: str,
 ):
     """Ensure export start API creates new export job."""
     response = admin_api_client.post(
-        path=reverse("export-artist-start"),
+        path=export_url,
         data={
             "file_format": "csv",
         },
@@ -53,15 +67,29 @@ def test_export_api_creates_export_job(
         ),
     ],
 )
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            reverse("export-artist-start"),
+            id="Export url",
+        ),
+        pytest.param(
+            f"{reverse('artists-export')}",
+            id="Action url",
+        ),
+    ],
+)
 def test_export_api_filtering(
     admin_api_client: test.APIClient,
     filter_query: str,
     filter_name: str,
     filter_value: str,
+    export_url: str,
 ):
     """Ensure export start API passes filter kwargs correctly."""
     response = admin_api_client.post(
-        path=f"{reverse('export-artist-start')}?{filter_query}",
+        path=f"{export_url}?{filter_query}",
         data={
             "file_format": "csv",
         },
@@ -100,14 +128,28 @@ def test_export_api_filtering(
         ),
     ],
 )
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            reverse("export-artist-start"),
+            id="Export url",
+        ),
+        pytest.param(
+            f"{reverse('artists-export')}",
+            id="Action url",
+        ),
+    ],
+)
 def test_export_api_ordering(
     admin_api_client: test.APIClient,
     ordering_query: str,
     ordering_value: collections.abc.Sequence[str],
+    export_url: str,
 ):
     """Ensure export start API passes ordering correctly."""
     response = admin_api_client.post(
-        path=f"{reverse('export-artist-start')}?{ordering_query}",
+        path=f"{export_url}?{ordering_query}",
         data={
             "file_format": "csv",
         },
@@ -123,12 +165,26 @@ def test_export_api_ordering(
 
 
 @pytest.mark.django_db(transaction=True)
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            f"{reverse('export-artist-start')}?id=invalid_id",
+            id="Export url with invalid filter_kwargs",
+        ),
+        pytest.param(
+            f"{reverse('artists-export')}?id=invalid_id",
+            id="Action url with invalid filter_kwargs",
+        ),
+    ],
+)
 def test_export_api_create_export_job_with_invalid_filter_kwargs(
     admin_api_client: test.APIClient,
+    export_url: str,
 ):
     """Ensure export start API with invalid kwargs return an error."""
     response = admin_api_client.post(
-        path=f"{reverse('export-artist-start')}?id=invalid_id",
+        path=export_url,
         data={
             "file_format": "csv",
         },
@@ -138,12 +194,26 @@ def test_export_api_create_export_job_with_invalid_filter_kwargs(
 
 
 @pytest.mark.django_db(transaction=True)
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            f"{reverse('export-artist-start')}?ordering=invalid_id",
+            id="Export url with invalid ordering",
+        ),
+        pytest.param(
+            f"{reverse('artists-export')}?ordering=invalid_id",
+            id="Action url with invalid ordering",
+        ),
+    ],
+)
 def test_export_api_create_export_job_with_invalid_ordering(
     admin_api_client: test.APIClient,
+    export_url: str,
 ):
     """Ensure export start API with invalid ordering return an error."""
     response = admin_api_client.post(
-        path=f"{reverse('export-artist-start')}?ordering=invalid_id",
+        path=export_url,
         data={
             "file_format": "csv",
         },
@@ -156,14 +226,28 @@ def test_export_api_create_export_job_with_invalid_ordering(
 
 
 @pytest.mark.django_db(transaction=True)
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            "export-artist-detail",
+            id="Model url",
+        ),
+        pytest.param(
+            "export-jobs-detail",
+            id="General url",
+        ),
+    ],
+)
 def test_export_api_detail(
     admin_api_client: test.APIClient,
     artist_export_job: ExportJob,
+    export_url: str,
 ):
     """Ensure export detail API shows current export job status."""
     response = admin_api_client.get(
         path=reverse(
-            "export-artist-detail",
+            export_url,
             kwargs={"pk": artist_export_job.id},
         ),
     )
@@ -172,15 +256,29 @@ def test_export_api_detail(
 
 
 @pytest.mark.django_db(transaction=True)
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            "export-artist-detail",
+            id="Model url",
+        ),
+        pytest.param(
+            "export-jobs-detail",
+            id="General url",
+        ),
+    ],
+)
 def test_import_user_api_get_detail(
     user: User,
     admin_api_client: test.APIClient,
     artist_export_job: ExportJob,
+    export_url: str,
 ):
     """Ensure import detail api for user returns only users jobs."""
     response = admin_api_client.get(
         path=reverse(
-            "export-artist-detail",
+            export_url,
             kwargs={"pk": artist_export_job.id},
         ),
     )
@@ -190,7 +288,7 @@ def test_import_user_api_get_detail(
     artist_export_job.save()
     response = admin_api_client.get(
         path=reverse(
-            "export-artist-detail",
+            export_url,
             kwargs={"pk": artist_export_job.id},
         ),
     )
@@ -205,17 +303,31 @@ def test_import_user_api_get_detail(
         ExportJob.ExportStatus.EXPORTING,
     ],
 )
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            "export-artist-cancel",
+            id="Model url",
+        ),
+        pytest.param(
+            "export-jobs-cancel",
+            id="General url",
+        ),
+    ],
+)
 def test_export_api_cancel(
     admin_api_client: test.APIClient,
     artist_export_job: ExportJob,
     allowed_cancel_status: ExportJob.ExportStatus,
+    export_url: str,
 ):
     """Ensure that export canceled with allowed statuses."""
     artist_export_job.export_status = allowed_cancel_status
     artist_export_job.save()
     response = admin_api_client.post(
         path=reverse(
-            "export-artist-cancel",
+            export_url,
             kwargs={"pk": artist_export_job.pk},
         ),
     )
@@ -233,17 +345,31 @@ def test_export_api_cancel(
         ExportJob.ExportStatus.CANCELLED,
     ],
 )
+@pytest.mark.parametrize(
+    argnames="export_url",
+    argvalues=[
+        pytest.param(
+            "export-artist-cancel",
+            id="Model url",
+        ),
+        pytest.param(
+            "export-jobs-cancel",
+            id="General url",
+        ),
+    ],
+)
 def test_export_api_cancel_with_errors(
     admin_api_client: test.APIClient,
     artist_export_job: ExportJob,
     incorrect_job_status: ExportJob.ExportStatus,
+    export_url: str,
 ):
     """Ensure that export job with incorrect statuses cannot be canceled."""
     artist_export_job.export_status = incorrect_job_status
     artist_export_job.save()
     response = admin_api_client.post(
         path=reverse(
-            "export-artist-cancel",
+            export_url,
             kwargs={"pk": artist_export_job.pk},
         ),
     )